MCP_Calendar/requirements.txt
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
mcp==1.18.0
python-dotenv==1.0.1
requests==2.32.3
tzlocal==5.3.1